pyproject.toml
[project]
name = "pipelines"
version = "1.0.0"
requires-python = "~=3.12"
description = "Demo of local and remote pipelines for calling and evaluating LLMs using Kubeflow and Vertex AI."
authors = [{name = "Gleb Lukicov"}]
readme = "README.md"

dependencies = [
    "docker>=7.1.0",
    "google-cloud-aiplatform>=1.72.0",
    "kfp>=2.10.1",
    "typer>=0.13.1",
]

[dependency-groups]
dev = [
    "mypy>=1.13.0",
    "ruff>=0.7.4",
]

[tool.uv]
package = true

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "kfp.*"
ignore_missing_imports = true

[project.scripts]
build-local = "scripts.cli:build_container"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
